fixtures updated
